PYIC-9002: Move VcDetails to it's own file
