+++ di-ipv-evcs-stub/lambdas/src/domain/requests/postRequest.ts
@@ -1,3 +1,3 @@
-import { VcDetails } from "../sharedTypes";
+import { VcDetails } from "../vcDetails";
 
 export interface PostRequest extends VcDetails {}
