+++ di-ipv-evcs-stub/lambdas/src/domain/sharedTypes.ts
@@ -1,12 +1,3 @@
-import { VcState, VCProvenance } from "./enums";
-
-export interface VcDetails {
-  vc: string;
-  state: VcState;
-  metadata?: object;
-  provenance?: VCProvenance;
-}
-
 export type StoredIdentityResponse = {
   si: Omit<VCIncludingStateAndMetadata, "state">;
   vcs: Omit<VCIncludingStateAndMetadata, "unsignedVot">[];
