+++ di-ipv-evcs-stub/lambdas/src/domain/vcDetails.ts
@@ -1,0 +1,8 @@
+import { VCProvenance, VcState } from "./enums";
+
+export interface VcDetails {
+  vc: string;
+  state: VcState;
+  metadata?: object;
+  provenance?: VCProvenance;
+}
